Phase-02/backend/requirements.txt
fastapi>=0.128.0
psycopg[binary]>=3.3.2
sqlmodel>=0.0.31
uvicorn>=0.40.0
python-dotenv>=1.0.0
pytest>=9.0.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
